igdb_puller/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "igdb-puller"
version = "0.1.0"
description = "Modular IGDB table downloader with CSV/Parquet/NDJSON export"
requires-python = ">=3.9"
authors = [{ name = "ProDoorknob" }]
readme = "README.md"
dependencies = [
  "requests>=2.31",
  "pandas>=2.1",
  "python-dotenv>=1.0",
  "pyarrow>=16.0",
]

[project.scripts]
igdb-puller = "igdb_puller.cli:main"